Changed -fit BG spectrum -Improve plotting -fit the BG channels
fit BG spectrum:
-To improve our fitting I added a spectrum from one of our BG sources
(channel 0) to the fitting routine.
-This requires to have ch_0.root and ch_1.root in the daqana folder

Improve plotting:
-I modified the PLOT_ON_SCREEN option such that we are also able to see
a histogram of the residuals, the distribution of the residuals and the
spectrum full spectrum

fit the BG channels:
-To see if our BG PMT's really measure the same spectrum every time I've
added a routine to fit the BG spectrum using the template ch_0.root
(which has to be in the daqana folder) to get a of this source (and see
how constant it is).