Added a settings.cfg file and read it in ecal.C
This contains per-institution settings for the calibration routine:
 * Number of peaks to fit per channel
 * Fitting range for the first peak of each channel

The settings can be specified per institute, and also for specific time ranges.
If settings are missing for the institute ecal.C will take the defaults which are hard-coded,
so if you don't add your settings to the settings.cfg file it will continue to work as now.